refactor(ai-agent): extract mode_normalize from acp/agent.rs
