+++ crates/aionui-ai-agent/src/manager/acp/mode_normalize.rs
@@ -1,0 +1,223 @@
+use aionui_api_types::AgentMetadata;
+
+pub(super) fn normalize_requested_mode(metadata: &AgentMetadata, mode: &str) -> String {
+    let trimmed = mode.trim();
+    if trimmed.is_empty() {
+        return String::new();
+    }
+
+    // AionUi persists the legacy aliases `yolo` / `yoloNoSandbox` while
+    // ACP backends expect their native mode id (e.g. `full-access` for
+    // Codex). Resolution is data-driven: the mapping lives on each
+    // catalog row's top-level `yolo_id` column. Backends without a
+    // `yolo_id` have no equivalent, so the alias passes through
+    // unchanged and `session/set_mode` gets the caller's original
+    // value.
+    if matches!(trimmed, "yolo" | "yoloNoSandbox")
+        && let Some(native) = metadata.yolo_id.as_deref()
+    {
+        return native.to_owned();
+    }
+
+    // Codex has legacy `default`/`autoEdit` aliases that map to its
+    // native `auto` mode. Keep the mapping data-driven by keying on the
+    // vendor backend label rather than re-introducing an AcpBackend
+    // enum variant.
+    if matches!(metadata.backend.as_deref(), Some("codex")) && matches!(trimmed, "default" | "autoEdit") {
+        return "auto".to_owned();
+    }
+
+    trimmed.to_owned()
+}
+
+/// Whether the agent described by `metadata` uses Claude-style meta resume
+/// (`session/new` with `_meta.claudeCode.options.resume`) instead of the
+/// generic `session/load` path.
+///
+/// Mirrors the AionUi frontend rule
+/// `useClaudeMetaResume = backend === 'claude' || !!caps?._meta?.claudeCode`.
+///
+/// Handshake blobs persisted by the backend are normalised to snake_case
+/// (see `sdk_to_snake_value`), so the lookup prefers `claude_code` and
+/// falls back to `claudeCode` for any blob that bypassed normalisation.
+pub(super) fn agent_metadata_uses_claude_meta_resume(metadata: &AgentMetadata) -> bool {
+    if metadata.backend.as_deref() == Some("claude") {
+        return true;
+    }
+    metadata
+        .handshake
+        .agent_capabilities
+        .as_ref()
+        .and_then(|caps| caps.get("_meta"))
+        .and_then(|meta| meta.get("claude_code").or_else(|| meta.get("claudeCode")))
+        .is_some()
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use aionui_api_types::AgentHandshake;
+    use aionui_common::AgentType;
+    use serde_json::json;
+
+    fn metadata_with_yolo_id(yolo_id: Option<&str>) -> AgentMetadata {
+        use aionui_api_types::{AgentSource, AgentSourceInfo, BehaviorPolicy};
+        AgentMetadata {
+            id: "test".into(),
+            icon: None,
+            name: "Test".into(),
+            name_i18n: None,
+            description: None,
+            description_i18n: None,
+            backend: None,
+            agent_type: AgentType::Acp,
+            agent_source: AgentSource::Builtin,
+            agent_source_info: AgentSourceInfo::default(),
+            enabled: true,
+            available: true,
+            command: None,
+            resolved_command: None,
+            args: vec![],
+            env: vec![],
+            native_skills_dirs: None,
+            behavior_policy: BehaviorPolicy::default(),
+            yolo_id: yolo_id.map(ToOwned::to_owned),
+            sort_order: 3130,
+            handshake: AgentHandshake::default(),
+        }
+    }
+
+    #[test]
+    fn normalize_requested_mode_rewrites_yolo_when_behavior_policy_maps_it() {
+        let meta = metadata_with_yolo_id(Some("full-access"));
+        assert_eq!(normalize_requested_mode(&meta, "yolo"), "full-access");
+        assert_eq!(normalize_requested_mode(&meta, "yoloNoSandbox"), "full-access");
+    }
+
+    #[test]
+    fn normalize_requested_mode_passes_through_when_no_yolo_id() {
+        let meta = metadata_with_yolo_id(None);
+        // No mapping configured — aliases flow through unchanged.
+        assert_eq!(normalize_requested_mode(&meta, "yolo"), "yolo");
+        assert_eq!(normalize_requested_mode(&meta, "yoloNoSandbox"), "yoloNoSandbox");
+    }
+
+    #[test]
+    fn normalize_requested_mode_passes_through_non_yolo_modes() {
+        let meta = metadata_with_yolo_id(Some("full-access"));
+        assert_eq!(normalize_requested_mode(&meta, "default"), "default");
+        assert_eq!(normalize_requested_mode(&meta, "read-only"), "read-only");
+        assert_eq!(
+            normalize_requested_mode(&meta, "bypassPermissions"),
+            "bypassPermissions"
+        );
+    }
+
+    /// Vendor-specific yolo rewrites are entirely data-driven by
+    /// `metadata.yolo_id`. Rebuild fixtures with the seed values
+    /// `006_agent_metadata.sql` would hydrate, then assert both yolo
+    /// aliases hit the native mode id for each vendor.
+    #[test]
+    fn normalize_requested_mode_rewrites_yolo_for_builtin_vendors() {
+        // Claude / Codebuddy → bypassPermissions.
+        let claude_like = metadata_with_yolo_id(Some("bypassPermissions"));
+        assert_eq!(normalize_requested_mode(&claude_like, "yolo"), "bypassPermissions");
+        assert_eq!(
+            normalize_requested_mode(&claude_like, "yoloNoSandbox"),
+            "bypassPermissions"
+        );
+        // Opencode → build.
+        let opencode_like = metadata_with_yolo_id(Some("build"));
+        assert_eq!(normalize_requested_mode(&opencode_like, "yolo"), "build");
+        // Cursor → agent.
+        let cursor_like = metadata_with_yolo_id(Some("agent"));
+        assert_eq!(normalize_requested_mode(&cursor_like, "yolo"), "agent");
+        // When a row has no yolo_id the alias flows through unchanged.
+        let gemini_like = metadata_with_yolo_id(None);
+        assert_eq!(normalize_requested_mode(&gemini_like, "yolo"), "yolo");
+    }
+
+    /// Codex's legacy `default` / `autoEdit` aliases should rewrite to
+    /// its native `auto` mode when the row's backend label is "codex".
+    /// Other backends must leave `default` / `autoEdit` untouched.
+    #[test]
+    fn normalize_requested_mode_rewrites_codex_default_and_auto_edit() {
+        let mut codex_meta = metadata_with_yolo_id(Some("full-access"));
+        codex_meta.backend = Some("codex".into());
+        assert_eq!(normalize_requested_mode(&codex_meta, "default"), "auto");
+        assert_eq!(normalize_requested_mode(&codex_meta, "autoEdit"), "auto");
+
+        let other = metadata_with_yolo_id(None);
+        assert_eq!(normalize_requested_mode(&other, "default"), "default");
+        assert_eq!(normalize_requested_mode(&other, "autoEdit"), "autoEdit");
+    }
+
+    /// Claude backend must take the `session/new` + `_meta.claudeCode.options.resume`
+    /// path so `mcpServers` are re-injected on resume. `backend == "claude"`
+    /// alone is enough — we don't need the handshake to advertise `_meta`.
+    #[test]
+    fn uses_claude_meta_resume_true_for_claude_backend() {
+        let mut meta = metadata_with_yolo_id(None);
+        meta.backend = Some("claude".into());
+        assert!(agent_metadata_uses_claude_meta_resume(&meta));
+    }
+
+    /// A non-Claude-labelled backend that still advertises
+    /// `agent_capabilities._meta.claudeCode` (snake_case, as persisted by
+    /// `sdk_to_snake_value`) must also follow the Claude resume path —
+    /// this matches the frontend's `!!caps?._meta?.claudeCode` check.
+    #[test]
+    fn uses_claude_meta_resume_true_for_meta_claude_code() {
+        let mut meta = metadata_with_yolo_id(None);
+        meta.backend = Some("custom-claude-wrapper".into());
+        meta.handshake.agent_capabilities = Some(json!({
+            "_meta": {
+                "claude_code": { "some": "flag" }
+            }
+        }));
+        assert!(agent_metadata_uses_claude_meta_resume(&meta));
+
+        // A handshake that bypassed snake_case normalisation (camelCase
+        // `claudeCode`) must still be recognised.
+        let mut camel_meta = metadata_with_yolo_id(None);
+        camel_meta.backend = Some("custom-claude-wrapper".into());
+        camel_meta.handshake.agent_capabilities = Some(json!({
+            "_meta": {
+                "claudeCode": { "some": "flag" }
+            }
+        }));
+        assert!(agent_metadata_uses_claude_meta_resume(&camel_meta));
+    }
+
+    /// Codex (and any non-Claude backend without the `_meta.claudeCode`
+    /// marker) must fall through to the `session/load` branch.
+    #[test]
+    fn uses_claude_meta_resume_false_for_codex() {
+        let mut meta = metadata_with_yolo_id(Some("full-access"));
+        meta.backend = Some("codex".into());
+        assert!(!agent_metadata_uses_claude_meta_resume(&meta));
+
+        // Codex with unrelated capability keys must still be false.
+        meta.handshake.agent_capabilities = Some(json!({
+            "load_session": true,
+            "_meta": { "codex": { "whatever": true } }
+        }));
+        assert!(!agent_metadata_uses_claude_meta_resume(&meta));
+    }
+
+    /// Metadata with no `backend` label and no handshake capabilities
+    /// must not opt into the Claude resume path.
+    #[test]
+    fn uses_claude_meta_resume_false_for_empty() {
+        let meta = metadata_with_yolo_id(None);
+        assert!(meta.backend.is_none());
+        assert!(meta.handshake.agent_capabilities.is_none());
+        assert!(!agent_metadata_uses_claude_meta_resume(&meta));
+    }
+
+    #[test]
+    fn normalize_requested_mode_trims_and_returns_empty_for_blank() {
+        let meta = metadata_with_yolo_id(Some("full-access"));
+        assert_eq!(normalize_requested_mode(&meta, "   "), "");
+    }
+}
